Translate variable/function names and comments within storage.js into English.
Update all references to the previous variable/function names elsewhere in the code to reflect this change. This includes a rename of the loadGroup and saveGroup function names into plural form, to make purpose clearer. Note that as part of this translation, the storage key for the active group name has been renamed, which could have a minor impact on backwards compatibility.